requirements.txt
kaggle==1.6.6
matplotlib==3.7.5
reportlab==4.1.0

apache-airflow-providers-google==10.17.0

numpy==1.24.4
scikit-learn==1.3.2
statsmodels==0.14.1
seaborn==0.13.2